# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'parking-be'","description":"'Parking Back End'","contact":{"name":"Hernan Jesus Morais","email":"hmorais@frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Parking Back End'"}],"paths":{"/lots/{lotId}/exit":{"put":{"tags":["lot-controller"],"operationId":"exitTrace","parameters":[{"name":"lotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"exit_date_time","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LotTrace"}}}}}}},"/lots/{lotId}/sale":{"post":{"tags":["lot-controller"],"operationId":"saleLot","parameters":[{"name":"lotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"entry_date_time","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LotTrace"}}}}}}},"/lots/{lotId}/rent":{"post":{"tags":["lot-controller"],"operationId":"rentLot","parameters":[{"name":"lotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"entry_date_time","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"months","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LotTrace"}}}}}}},"/lots/{lotId}/entry":{"post":{"tags":["lot-controller"],"operationId":"entryLot","parameters":[{"name":"lotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"entry_date_time","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LotTrace"}}}}}}},"/vehicles":{"get":{"tags":["vehicle-wrapper-controller"],"operationId":"getList","parameters":[{"name":"patente","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/lots":{"get":{"tags":["lot-controller"],"operationId":"getList_1","parameters":[{"name":"floor","in":"query","required":false,"schema":{"type":"string","enum":["SUBSUELO","PLANTA BAJA","PRIMER PISO"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lot"}}}}}}}},"/lots/{lotId}":{"get":{"tags":["lot-controller"],"operationId":"getById","parameters":[{"name":"lotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Lot"}}}}}}}},"components":{"schemas":{"LotPrice":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["ALQUILER MENSUAL","TEMPORARIO","PRIVADO"]},"section":{"type":"string","enum":["AUTOS","CAMIONETAS","MOTOS"]},"price":{"type":"number"},"active":{"type":"boolean"},"valid_from":{"type":"string","format":"date-time"},"valid_to":{"type":"string","format":"date-time"}}},"LotTrace":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"vehicle":{"$ref":"#/components/schemas/Vehicle"},"amount":{"type":"number"},"lotPrice":{"$ref":"#/components/schemas/LotPrice"},"date_from":{"type":"string","format":"date-time"},"date_to":{"type":"string","format":"date-time"}}},"Vehicle":{"type":"object","properties":{"id":{"type":"string"},"brand":{"type":"string"},"model":{"type":"string"},"color":{"type":"string"},"type":{"type":"string"}}},"Lot":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"floor":{"type":"string","enum":["SUBSUELO","PLANTA BAJA","PRIMER PISO"]},"section":{"type":"string","enum":["AUTOS","CAMIONETAS","MOTOS"]},"type":{"type":"string","enum":["ALQUILER MENSUAL","TEMPORARIO","PRIVADO"]},"status":{"type":"string","enum":["LIBRE","OCUPADO"]},"occupied_by":{"$ref":"#/components/schemas/LotTrace"}}}}}}