# src/main/java/ar/edu/utn/frc/tup/lc/iv/models/LotPrice.java
package ar.edu.utn.frc.tup.lc.iv.models;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LotPrice {

    /**
     IMPORTANTE: No modificar esta clase. NO ES NECESARIO
     */

    private Long id;
    private LotType type;
    private Section section;
    private BigDecimal price;
    @JsonProperty("valid_from")
    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
    private LocalDateTime validFrom;
    @JsonProperty("valid_to")
    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
    private LocalDateTime validTo;
    private Boolean active;
}